prescription-reader/requirements.txt
flask
flask-cors
opencv-python
pytesseract
Pillow
numpy